test(data-loader): refactor tests
