+++ packages/data-loader/src/kintone/__tests__/download/fixtures/can_get_records.ts
@@ -1,5 +1,5 @@
-import { RecordForExport } from "../../../types/data-loader";
-import { KintoneRecordForResponse } from "../../../types/kintone";
+import { RecordForExport } from "../../../../types/data-loader";
+import { KintoneRecordForResponse } from "../../../../types/kintone";
 
 export const input: KintoneRecordForResponse[] = [
   {
