+++ packages/data-loader/src/kintone/__tests__/upload/fixtures/update_records/input.ts
@@ -1,0 +1,26 @@
+import { RecordForImport } from "../../../../../types/data-loader/import/record";
+
+export const input: RecordForImport[] = [
+  {
+    singleLineText: {
+      value: "value1",
+    },
+    number: {
+      value: "1",
+    },
+    date: {
+      value: "2022-03-01",
+    },
+  },
+  {
+    singleLineText: {
+      value: "value2",
+    },
+    number: {
+      value: "2",
+    },
+    date: {
+      value: "2022-04-01",
+    },
+  },
+];
