+++ packages/data-loader/src/kintone/__tests__/upload/fixtures/update_records/properties.ts
@@ -1,0 +1,55 @@
+import { KintoneFormFieldProperty } from "@kintone/rest-api-client";
+
+export const properties: Record<string, KintoneFormFieldProperty.OneOf> = {
+  singleLineText: {
+    type: "SINGLE_LINE_TEXT",
+    code: "singleLineText",
+    label: "singleLineText",
+    noLabel: false,
+    required: false,
+    minLength: "",
+    maxLength: "",
+    expression: "",
+    hideExpression: false,
+    unique: true,
+    defaultValue: "",
+  },
+  number: {
+    type: "NUMBER",
+    code: "number",
+    label: "number",
+    noLabel: false,
+    required: true,
+    minValue: "",
+    maxValue: "",
+    digit: false,
+    unique: true,
+    defaultValue: "",
+    displayScale: "",
+    unit: "",
+    unitPosition: "BEFORE",
+  },
+  singleLineText_nonUnique: {
+    type: "SINGLE_LINE_TEXT",
+    code: "singleLineText",
+    label: "singleLineText",
+    noLabel: false,
+    required: false,
+    minLength: "",
+    maxLength: "",
+    expression: "",
+    hideExpression: false,
+    unique: false,
+    defaultValue: "",
+  },
+  date: {
+    type: "DATE",
+    code: "date",
+    label: "date",
+    noLabel: false,
+    required: false,
+    unique: false,
+    defaultValue: "",
+    defaultNowValue: true,
+  },
+};
